mcp-server/requirements.txt
# MCP Server
mcp>=0.9.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Date parsing
python-dateutil>=2.8.2

# HTTP server (for health checks)
aiohttp>=3.9.0

# Environment variables
python-dotenv>=1.0.0
